update for 1007, move jsf tests to ejb30.lite.packaging.war.xmloverride.ejbrefjsf
