+++ install/jakartaee/other/vehicle.properties
@@ -303,8 +303,8 @@ com/sun/ts/tests/ejb30/lite/packaging/war/webinflib/JsfClient.java = ejblitejsf
 com/sun/ts/tests/ejb30/lite/packaging/war/webinflib/Client.java = ejbliteservlet ejbliteservlet2 ejblitejsp
 com/sun/ts/tests/ejb30/lite/packaging/war/jsfwebinflibonly = ejblitejsf
 com/sun/ts/tests/ejb30/lite/packaging/war/webinflibonly = ejbliteservlet ejbliteservlet2 ejblitejsp
-com/sun/ts/tests/ejb30/lite/packaging/war/xmloverride/ejbref/JsfClient.java = ejblitejsf
-com/sun/ts/tests/ejb30/lite/packaging/war/xmloverride/ejbref/Client.java = ejbliteservlet ejbliteservlet2 ejblitejsp
+com/sun/ts/tests/ejb30/lite/packaging/war/xmloverride/ejbrefjsf = ejblitejsf
+com/sun/ts/tests/ejb30/lite/packaging/war/xmloverride/ejbref = ejbliteservlet ejbliteservlet2 ejblitejsp
 
 com/sun/ts/tests/ejb30/lite/naming/context/Client.java = ejbliteservlet ejbliteservlet2 ejblitejsp ejbembed
 com/sun/ts/tests/ejb30/lite/naming/context/Client.java#lookup = ejbliteservlet ejbliteservlet2 ejblitejsp
