+++ src/com/sun/ts/tests/ejb30/lite/packaging/war/xmloverride/ejbrefjsf/build.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <!--
 
-    Copyright (c) 2018,2022 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2022 Oracle and/or its affiliates. All rights reserved.
 
     This program and the accompanying materials are made available under the
     terms of the Eclipse Public License v. 2.0, which is available at
@@ -25,6 +25,7 @@
   <target name="package">
     <ts.vehicles name="${app.name}" buildleveloverride="2"
       classes="
+        com/sun/ts/tests/ejb30/common/lite/EJBLiteJsfClientBase.class,
         com/sun/ts/tests/ejb30/common/helper/ServiceLocator.class,
         com/sun/ts/tests/ejb30/lite/xmloverride/ejbref/XmlOverrideBean.class,
         com/sun/ts/tests/ejb30/lite/xmloverride/ejbref/XmlOverride2Bean.class
